+++ packages/export-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/export-web
 
+## 0.43.1-pre.0
+
+### Patch Changes
+
+- @glissade/backend-canvas2d@0.43.1-pre.0
+- @glissade/core@0.43.1-pre.0
+- @glissade/scene@0.43.1-pre.0
+
 ## 0.43.0
 
 ### Patch Changes