+++ packages/narrate/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @glissade/narrate
 
+## 0.43.1-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.43.1-pre.0
+- @glissade/scene@0.43.1-pre.0
+
 ## 0.43.0
 
 ### Patch Changes