+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/react
 
+## 0.43.1-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.43.1-pre.0
+- @glissade/player@0.43.1-pre.0
+- @glissade/scene@0.43.1-pre.0
+
 ## 0.43.0
 
 ### Patch Changes