+++ packages/sfx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/sfx
 
+## 0.43.1-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.43.1-pre.0
+
 ## 0.43.0
 
 ### Patch Changes